Refactor and extend test suite.
