+++ src/_test_model_checking/mod.rs
@@ -1,0 +1,72 @@
+/// **(internal)** Test evaluating several important formulae on several models.
+/// Compare result numbers with the expected numbers (precomputed or obtained elsewhere).
+mod _test_against_precomputed;
+
+/// **(internal)**  Test evaluation of pairs of equivalent formulae on several models.
+mod _test_formula_equivalences;
+
+/// **(internal)**  Test evaluation of pairs of equivalent `formula patterns` on several models.
+/// Try evaluating several versions of the pattern formula (by substituting random expressions).
+mod _test_pattern_equivalences;
+
+/// **(internal)**  Test evaluation of extended formulae, where special propositions
+/// are evaluated as various simple pre-computed sets.
+mod _test_extended_formulae;
+
+/// Test that the (potentially unsafe) optimization on cases where it should be applicable.
+mod _test_unsafe_optimization;
+
+/// **(internal)**  Test evaluation of formulae where quantified variables are given
+/// restricted domains.
+mod _test_variable_domains;
+
+/// **(internal)**  Utilities used in tests, such as generating sets of random formula trees.
+mod _test_util;
+
+// model FISSION-YEAST-2008
+const MODEL_YEAST: &str = r"
+targets,factors
+Cdc25, ((!Cdc2_Cdc13 & (Cdc25 & !PP)) | ((Cdc2_Cdc13 & (!Cdc25 & !PP)) | (Cdc2_Cdc13 & Cdc25)))
+Cdc2_Cdc13, (!Ste9 & (!Rum1 & !Slp1))
+Cdc2_Cdc13_A, (!Ste9 & (!Rum1 & (!Slp1 & (!Wee1_Mik1 & Cdc25))))
+PP, Slp1
+Rum1, ((!SK & (!Cdc2_Cdc13 & (!Rum1 & (!Cdc2_Cdc13_A & PP)))) | ((!SK & (!Cdc2_Cdc13 & (Rum1 & !Cdc2_Cdc13_A))) | ((!SK & (!Cdc2_Cdc13 & (Rum1 & (Cdc2_Cdc13_A & PP)))) | ((!SK & (Cdc2_Cdc13 & (Rum1 & (!Cdc2_Cdc13_A & PP)))) | (SK & (!Cdc2_Cdc13 & (Rum1 & (!Cdc2_Cdc13_A & PP))))))))
+SK, Start
+Slp1, Cdc2_Cdc13_A
+Start, false
+Ste9, ((!SK & (!Cdc2_Cdc13 & (!Ste9 & (!Cdc2_Cdc13_A & PP)))) | ((!SK & (!Cdc2_Cdc13 & (Ste9 & !Cdc2_Cdc13_A))) | ((!SK & (!Cdc2_Cdc13 & (Ste9 & (Cdc2_Cdc13_A & PP)))) | ((!SK & (Cdc2_Cdc13 & (Ste9 & (!Cdc2_Cdc13_A & PP)))) | (SK & (!Cdc2_Cdc13 & (Ste9 & (!Cdc2_Cdc13_A & PP))))))))
+Wee1_Mik1, ((!Cdc2_Cdc13 & (!Wee1_Mik1 & PP)) | ((!Cdc2_Cdc13 & Wee1_Mik1) | (Cdc2_Cdc13 & (Wee1_Mik1 & PP))))
+";
+
+// model MAMMALIAN-CELL-CYCLE-2006
+const MODEL_CELL_CYCLE: &str = r"
+targets,factors
+v_Cdc20, v_CycB
+v_Cdh1, ((v_Cdc20 | (v_p27 & !v_CycB)) | !(((v_p27 | v_CycB) | v_CycA) | v_Cdc20))
+v_CycA, ((v_CycA & !(((v_Cdh1 & v_UbcH10) | v_Cdc20) | v_Rb)) | (v_E2F & !(((v_Cdh1 & v_UbcH10) | v_Cdc20) | v_Rb)))
+v_CycB, !(v_Cdc20 | v_Cdh1)
+v_CycE, (v_E2F & !v_Rb)
+v_E2F, ((v_p27 & !(v_CycB | v_Rb)) | !(((v_p27 | v_Rb) | v_CycB) | v_CycA))
+v_Rb, ((v_p27 & !(v_CycD | v_CycB)) | !((((v_CycE | v_p27) | v_CycB) | v_CycD) | v_CycA))
+v_UbcH10, (((((v_UbcH10 & ((v_Cdh1 & ((v_CycB | v_Cdc20) | v_CycA)) | !v_Cdh1)) | (v_CycA & !v_Cdh1)) | (v_Cdc20 & !v_Cdh1)) | (v_CycB & !v_Cdh1)) | !((((v_UbcH10 | v_Cdh1) | v_CycB) | v_Cdc20) | v_CycA))
+v_p27, ((v_p27 & !((v_CycD | (v_CycA & v_CycE)) | v_CycB)) | !((((v_CycE | v_p27) | v_CycB) | v_CycD) | v_CycA))
+";
+
+// largely parametrized version of the model ASYMMETRIC-CELL-DIVISION-B
+const MODEL_CELL_DIVISION: &str = r"
+DivJ -?? DivK
+PleC -?? DivK
+DivK -?? DivL
+DivL -?? CckA
+CckA -?? ChpT
+ChpT -?? CpdR
+CpdR -?? ClpXP_RcdA
+ChpT -?? CtrAb
+ClpXP_RcdA -?? CtrAb
+DivK -?? DivJ
+PleC -?? DivJ
+DivK -?? PleC
+$CckA: DivL
+$ChpT: CckA
+$DivK: (!PleC & DivJ)
+";
