+++ src/evaluation/mark_duplicate_subform.rs
@@ -185,7 +185,7 @@ mod tests {
 
     #[test]
     /// Compare automatically detected duplicate sub-formulae to expected ones.
-    fn test_duplicates_single_simple() {
+    fn duplicates_single_simple() {
         let formula = "!{x}: 3{y}: (AX {x} & AX {y})";
         let expected_duplicates = HashMap::from([("(AX {var0})".to_string(), 1)]);
 
@@ -201,7 +201,7 @@ mod tests {
 
     #[test]
     /// Compare automatically detected duplicate sub-formulae to expected ones.
-    fn test_duplicates_single_complex() {
+    fn duplicates_single_complex() {
         let formula = "(!{x}: 3{y}: ((AG EF {x} & AG EF {y}) & (EF {y}))) & (!{z}: EF {z})";
         let expected_duplicates = HashMap::from([
             ("(AG (EF {var0}))".to_string(), 1),
@@ -220,7 +220,7 @@ mod tests {
     #[test]
     /// Compare automatically detected duplicate sub-formulae to expected ones.
     /// Use multiple input formulae.
-    fn test_duplicates_multiple() {
+    fn duplicates_multiple_formulae() {
         let formulae = vec![
             "!{x}: 3{y}: (AX {x} & AX {y})",
             "!{x}: (AX {x})",
@@ -247,7 +247,7 @@ mod tests {
 
     #[test]
     /// Test that wild-card propositions are also detected correctly (opposed to other terminals).
-    fn test_duplicates_wild_cards() {
+    fn duplicates_with_wild_cards() {
         // define a placeholder bn
         let bn = BooleanNetwork::try_from_bnet("v1, v1").unwrap();
         let ctx = SymbolicContext::new(&bn).unwrap();
@@ -262,7 +262,7 @@ mod tests {
 
     #[test]
     /// Test that duplicates with same duplicates are detected, but with different domains not.
-    fn test_duplicates_domains() {
+    fn duplicates_with_domains() {
         // define a placeholder bn
         let bn = BooleanNetwork::try_from_bnet("v1, v1").unwrap();
         let ctx = SymbolicContext::new(&bn).unwrap();
